Revert MIRV attacks enemy units (#1452)
This reverts commit 185fcb5fb0e65c309f1a7640772b3c752de3901f.

## Description:

This change made the MIRV way too powerful. If two players MIRV each
other it creates a stalemate because it takes so long to recover.

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I understand that submitting code with bugs that could have been
caught through manual testing blocks releases and new features for all
contributors

## Please put your Discord username so you can be contacted if a bug or
regression is found:

evan