Fix .github/workflows/pr-description.yml to support current PR structure (#1506)
## Description:

This PR fixes an issue in .github/workflows/pr-description.yml where the
logic for detecting checkboxes did not match the current PR format.
The condition was incorrect, so it has been updated to correctly
validate the checkbox state.

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I have read and accepted the CLA aggreement (only required once).

## Please put your Discord username so you can be contacted if a bug or
regression is found:

aotumuri